bigquery: add client.list_dataset_tables
Remove Dataset.list_tables
